# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Textile's HTTP REST API Documentation",
        "title": "Textile REST API",
        "termsOfService": "https://github.com/b582q9/go-textile-sapien/blob/master/TERMS",
        "contact": {
            "name": "Textile",
            "url": "https://textile.io/",
            "email": "contact@textile.io"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/b582q9/go-textile-sapien/blob/master/LICENSE"
        },
        "version": "0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/messages": {
            "get": {
                "description": "Paginates thread messages",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Paginates thread messages",
                "parameters": [
                    {
                        "type": "string",
                        "default": "thread=,offset=,limit=10",
                        "description": "thread: Thread ID (can also use 'default', omit for all), offset: Offset ID to start listing from (omit for latest), limit: List page size (default: 5)",
                        "name": "X-Textile-Opts",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "messages",
                        "schema": {
                            "$ref": "#/definitions/pb.TextList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/messages/{block}": {
            "get": {
                "description": "Gets a thread message by block ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "messages"
                ],
                "summary": "Get thread message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "block id",
                        "name": "block",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "$ref": "#/definitions/pb.Text"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/threads/{id}/messages": {
            "post": {
                "description": "Adds a message to a thread",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "threads"
                ],
                "summary": "Add a message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "urlescaped message body",
                        "name": "X-Textile-Args",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message",
                        "schema": {
                            "$ref": "#/definitions/pb.Text"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.Comment": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "target": {
                    "type": "object",
                    "$ref": "#/definitions/pb.FeedItem"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/pb.User"
                }
            }
        },
...
        "pb.FeedItem": {
            "type": "object",
            "properties": {
                "block": {
                    "type": "string"
                },
                "payload": {
                    "type": "string"
                },
                "thread": {
                    "type": "string"
                }
            }
        },
...
        "pb.Like": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "target": {
                    "type": "object",
                    "$ref": "#/definitions/pb.FeedItem"
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/pb.User"
                }
            }
        },
...
        "pb.Text": {
            "type": "object",
            "properties": {
                "block": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Comment"
                    }
                },
                "date": {
                    "type": "string"
                },
                "likes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Like"
                    }
                },
                "user": {
                    "type": "object",
                    "$ref": "#/definitions/pb.User"
                }
            }
        },
        "pb.TextList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Text"
                    }
                }
            }
        },
...
        "pb.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}